+++ spec/models/carto/api_key_spec.rb
@@ -19,6 +19,7 @@ describe Carto::ApiKey do
   end
 
   def database_grant(database_schema = 'wadus', table_name = 'wadus',
+                     owner = false,
                      permissions: ['insert', 'select', 'update', 'delete'],
                      schema_permissions: ['create'])
     {
@@ -27,6 +28,7 @@ describe Carto::ApiKey do
         {
           schema: database_schema,
           name: table_name,
+          owner: owner,
           permissions: permissions
         }
       ],
@@ -39,7 +41,7 @@ describe Carto::ApiKey do
     }
   end
 
-  def table_grant(database_schema = 'wadus', table_name = 'wadus',
+  def table_grant(database_schema = 'wadus', table_name = 'wadus', owner = false,
                   permissions: ['insert', 'select', 'update', 'delete'])
     {
       type: "database",
@@ -47,6 +49,7 @@ describe Carto::ApiKey do
         {
           schema: database_schema,
           name: table_name,
+          owner: owner,
           permissions: permissions
         }
       ]
@@ -166,6 +169,40 @@ describe Carto::ApiKey do
       end
     end
 
+    it 'master role is able to drop the table created by regular api with schema grants' do
+      grants = [schema_grant(@carto_user1.database_schema), apis_grant]
+      api_key = @carto_user1.api_keys.create_regular_key!(name: 'drop_by_master', grants: grants)
+
+      with_connection_from_api_key(api_key) do |connection|
+        connection.execute("create table \"#{@carto_user1.database_schema}\".test_table as select 1 as test")
+        connection.execute("select count(1) from \"#{@carto_user1.database_schema}\".test_table") do |result|
+          result[0]['count'].should eq '1'
+        end
+      end
+
+      @carto_user1.in_database.execute("drop table test_table")
+    end
+
+    it 'reassign created table ownership after delete the api key' do
+      grants = [schema_grant(@carto_user1.database_schema), apis_grant]
+      api_key = @carto_user1.api_keys.create_regular_key!(name: 'drop_test', grants: grants)
+
+      with_connection_from_api_key(api_key) do |connection|
+        connection.execute("create table \"#{@carto_user1.database_schema}\".test_table as select 1 as test")
+        connection.execute("select count(1) from \"#{@carto_user1.database_schema}\".test_table") do |result|
+          result[0]['count'].should eq '1'
+        end
+      end
+
+      api_key.destroy
+
+      ownership_query = "select pg_catalog.pg_get_userbyid(relowner) as owner from pg_class where relname = 'test_table'"
+      @carto_user1.in_database.execute(ownership_query) do |result|
+        result[0]['owner'].should eq @carto_user1.database_username
+      end
+      @carto_user1.in_database.execute("drop table test_table")
+    end
+
     it 'fails to grant to a non-existent schema' do
       expect {
         grants = [schema_grant('not-exists'), apis_grant]
@@ -277,6 +314,26 @@ describe Carto::ApiKey do
 
       @user1.in_database.run(drop_query)
       api_key.destroy
+    end
+
+    it 'show ownership of the tables for the user' do
+      grants = [schema_grant(@carto_user1.database_schema), apis_grant]
+      api_key = @carto_user1.api_keys.create_regular_key!(name: 'table_owner_test', grants: grants)
+
+      with_connection_from_api_key(api_key) do |connection|
+        connection.execute("create table \"#{@carto_user1.database_schema}\".test_table as select 1 as test")
+        connection.execute("select count(1) from \"#{@carto_user1.database_schema}\".test_table") do |result|
+          result[0]['count'].should eq '1'
+        end
+      end
+
+      permissions = api_key.table_permissions_from_db
+
+      permissions.each do |p|
+        if p.name == 'test_table'
+          p.owner.should eq true
+        end
+      end
     end
 
     let (:grants) { [database_grant(@table1.database_schema, @table1.name), apis_grant] }