+++ NEWS.md
@@ -9,7 +9,8 @@ sudo make install
 ```
 
 ### Features
-- None yet
+- OAuth:
+  - Regular api keys are now able to create tables ([#14978](https://github.com/CartoDB/cartodb/issues/14978))
 
 ### Bug fixes / enhancements
 - Fixes migrations for users with OAuth related data (#14600)